2331147@info Chat/2331147@info Chat/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.32.0
readability-lxml>=0.8.1
lxml>=5.0.0
faiss-cpu>=1.8.0
lxml_html_clean>=0.1.0
sentence-transformers>=2.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.0.0
urllib3>=2.0.0
tqdm>=4.60.0
openai>=1.30.0
streamlit>=1.35.0
requests-cache>=1.0.0
